refactor: CarsValidator 로직 수정
- Validator는 입력 데이터의 적합성만 판단하는 기능만 가지는것이라고 생각함.
- Validator의 반환값을 void로 수정.
- 문자열로 들어온 이름들의 전처리,validate,리스트화를 시켜주는 CarsGenerator 클래스를 구성.
